requirements.txt
coverage==4.4.1
Django==1.11.4
django-autoslug==1.9.3
django-nose==1.4.5
prospector==0.12.7
